requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
pandas==2.1.3
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-genai==0.0.6
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
